requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9.0
django-cors-headers>=4.0.0
python-decouple>=3.8
